pyproject.toml
[tool.poetry]
name = "fastapi-project"
version = "0.1.0"
description = ""
authors = ["Isfireomat <dfdfgfgf666@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.115.2"}
alembic = "^1.13.3"
pillow = "^11.0.0"
psycopg2 = "^2.9.10"
sqlalchemy = "^2.0.36"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
asyncpg = "^0.30.0"
redis = {extras = ["async"], version = "^5.2.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
